% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_zaga_parameters}
\alias{extract_zaga_parameters}
\title{Extract Parameters from Fitted ZAGA Model}
\usage{
extract_zaga_parameters(model, p_adjust)
}
\arguments{
\item{model}{Fitted gamlss model object}

\item{p_adjust}{Adjusted p-value for critical value calculation}
}
\value{
Tibble with mu, sigma, nu, crit, and fillibens
}
\description{
Extract Parameters from Fitted ZAGA Model
}
\keyword{internal}
