% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{revcom}
\alias{revcom}
\title{Reverse Complement a DNA String}
\usage{
revcom(x)
}
\arguments{
\item{x}{A single \code{character} string representing a DNA
sequence (e.g., "ACGT").}
}
\value{
A single \code{character} string representing the
  reverse complement of the input sequence.
}
\description{
Calculates the reverse complement of a DNA character string.
}
\examples{
seq <- "ATGCCG"
MultiEditR:::revcom(seq)
# Expected output: "CGGCAT"
}
